move stats function for marathon to seperate file
